refactor(core): removes core module and moves the files inside to more relevant directories
