+++ src/app/main/declarations/dsp-api-tokens.ts
@@ -1,4 +1,4 @@
-import { InjectionToken, NgModule } from '@angular/core';
+import { InjectionToken } from '@angular/core';
 import { KnoraApiConfig, KnoraApiConnection } from '@dasch-swiss/dsp-js';
 
 // config for dsp-js-lib (@dasch-swiss/dsp-js) config object
@@ -6,12 +6,3 @@ export const DspApiConfigToken = new InjectionToken<KnoraApiConfig>('DSP api con
 
 // connection config for dsp-js-lib (@dasch-swiss/dsp-js) connection
 export const DspApiConnectionToken = new InjectionToken<KnoraApiConnection>('DSP api connection instance');
-
-
-@NgModule({
-    declarations: [],
-    imports: [],
-    exports: []
-})
-export class AppCoreModule {
-}
