+++ src/app/main/services/app-init.service.ts
